Refactor to remove Uploader in favor of RackspaceFileProvider
