+++ atst/domain/exceptions.py
@@ -30,3 +30,7 @@ class UnauthenticatedError(Exception):
     @property
     def message(self):
         return str(self)
+
+
+class UploadError(Exception):
+    pass
