+++ config/test.ini
@@ -3,3 +3,4 @@ ENVIRONMENT = test
 PGDATABASE = atat_test
 CRL_DIRECTORY = tests/fixtures/crl
 WTF_CSRF_ENABLED = false
+STORAGE_PROVIDER=LOCAL
